# lib/api/openapi.yaml
openapi: 3.0.0
info:
  title: AI-LinkMO User Demo, Open API
  version: 1.2.1
paths:
  /risk:
    get:
      summary: Get risk
      operationId: handlers.risks
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: false
          description: Unique identifier for the resource
        - in: query
          name: byod
          schema:
            type: boolean
          required: false
          description: Import your own data into the Ai Atlas Nexus knowledge graph
        - in: query
          name: isDefinedByTaxonomy
          schema:
            type: string
          required: false
          description: Filter by taxonomy
        - in: query
          name: hasDocumentation
          schema:
            type: string
          required: false
          description: Specify a document ID for the selected scope
        - in: query
          name: isPartOf
          schema:
            type: string
          required: false
          description: Filter by partof (some entity)
        - in: query
          name: hasPart
          schema:
            type: string
          required: false
          description: Filter by part
        - in: query
          name: detectsRiskConcept
          schema:
            type: string
          required: false
          description: Filter by concept that detects a risk
        - in: query
          name: descriptor
          schema:
            type: string
          required: false
          description: Filter by descriptor
        - in: query
          name: isDetectedBy
          schema:
            type: string
          required: false
          description: Filter by detection concepts
        - in: query
          name: risk_type
          schema:
            type: string
          required: false
          description: Specify type of selected scope
        - in: query
          name: implementedByAdapter
          schema:
            type: string
          required: false
          description: Filter by adapter that implements the risk
        - in: query
          name: phase
          schema:
            type: string
            enum: ["training-tuning", "inference"]
          required: false
          description: Filter by risk phase.
        - in: query
          name: related
          schema:
            type: boolean
          required: false
          description: Fetch related items for the given scope ID
        - in: query
          name: related_ids
          schema:
            type: boolean
          required: false
          description: Fetch IDs of related items for the given scope ID
      responses:
        "200":
          description: Risk response
          content:
            application/json:
              schema:
                type: object
...
  /ares:
    post:
      summary: Run ARES evaluation
      operationId: handlers.ares
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                risks:
                  type: array
                  items:
                    type: object
                    description: Risk object
                inference_engine:
                  type: string
                  description: Inference engine identifier
                target:
                  type: object
                  description: Target dictionary
      responses:
        "200":
          description: ARES evaluation completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    description: Evaluation result